+++ lua/commands.lua
@@ -5,7 +5,7 @@ local M = {}
 
 -- noahvim menu + command
 M.noahvim_menu = function()
-  require("noahvim.picker").create("Noahvim Menu", {
+  require("util.picker").create("Noahvim Menu", {
     { "Toggle File Explorer",  "NvimTreeToggle" },
     { "Noahvim Git Menu", "NoahvimGit" },
     { "Noahvim Help Page", "NoahvimHelp" },
@@ -21,7 +21,7 @@ vim.api.nvim_create_user_command("NoahvimMenu", M.noahvim_menu, { nargs = 0 })
 
 -- git menu + command
 M.git_menu = function()
-  require("noahvim.picker").create("Noahvim Git Menu", {
+  require("util.picker").create("Noahvim Git Menu", {
     { "Git Status", "Telescope git_status" },
     { "Git Commits", "Telescope git_commits" },
     { "Git Branch Commits", "Telescope git_bcommits" },
@@ -38,7 +38,7 @@ vim.api.nvim_create_user_command("NoahvimGit", M.git_menu, { nargs = 0 })
 
 -- terminal menu + command
 M.term_menu = function()
-  require("noahvim.picker").create("Noahvim Terminal Menu", {
+  require("util.picker").create("Noahvim Terminal Menu", {
     { "Open with Horizontal Split", "split | term" },
     { "Open with Vertical Split", "vsplit | term" }
   }, require("telescope.themes").get_ivy({
@@ -53,7 +53,7 @@ vim.api.nvim_create_user_command("NoahvimTerm", M.term_menu, { nargs = 0 })
 -- create window helper
 M.create_window_from_file = function(filename)
   local keymapsdoc = vim.api.nvim_get_runtime_file(filename, false)[1]
-  return require("noahvim.window").create_with_md_file(keymapsdoc)
+  return require("util.window").create({ filepath = keymapsdoc })
 end
 
 -- show help window + command
@@ -77,5 +77,7 @@ end
 
 vim.api.nvim_create_user_command("NoahvimCredits", M.credits_window, { nargs = 0 })
 
+-- notepad command
+vim.api.nvim_create_user_command("Notepad", require("ui.notepad").launch, { nargs = 0 })
 
 return M