+++ lua/config/tree.lua
@@ -9,9 +9,31 @@ local function on_attach(bufnr)
 
   -- default keybinds
   api.config.mappings.default_on_attach(bufnr)
+
+  -- file actions picker
+  local actions = function()
+    require("util.picker").create("Noahvim File Actions", {
+      { "Create node", api.fs.create },
+      { "Remove node", api.fs.remove },
+      { "Trash node", api.fs.trash },
+      { "Rename node", api.fs.rename },
+      { "Fully rename node", api.fs.rename_sub },
+      { "Copy", api.fs.copy.node },
+    }, require("telescope.themes").get_ivy({
+      layout_config = {
+        height = 10
+      }
+    }))
+  end
+
+  vim.keymap.set("n", "<C-Space>", actions, {
+    buffer = bufnr,
+    noremap = true,
+    silent = true,
+  })
 end
 
-M.config = function()
+M.setup = function()
   vim.g.loaded_netrw = 1
   vim.g.loaded_netrwPlugin = 1
 
@@ -57,8 +79,7 @@ end
 
 -- keybinds
 M.keybinds = {
-  { "n", "<leader>t", "<CMD>NvimTreeToggle<CR>", "[t]oggle: toggle file explorer" },
-  { "n", "<leader>tf", "<CMD>NvimTreeFocus<CR>", "[f]ocus: focus the file explorer"}
+  { "n", "<leader>t", "<CMD>NvimTreeToggle<CR>", "[t]oggle: toggle file explorer" }
 }
 
 return M